Create the viewset for the Garments display endpoint.
This viewset will handle all the CRUD functionalities given an appropriate enpoint and an HTTP method (this will be further handled by Django router).
Note that the "UD" operations are set to require an authentication. This can be done by creating a superuser through django manage.py and logging in on the 'admin/' endpoint.