Loading the massive JL file into the database #4: (major) create a utility script that PREPROCESSES and converts the S3 '.jl' to two JSON files compliant with the Django's "loaddata" procedure.
One JSON files is for all the garments and the other for the 'images' property (which is a sub-object in the original S3 file).
Both JSONs have the structure that is compliant with django's "python manage.py loaddata" method (see the docs for fixtures) and the model structure we defined in commit #1 of this process.
JSONs are then saved in 'fixture's directory.
I will NOT commit the created JSONS again due to confidentiality and size reasons so REMEMBER to run this utility if reproducing the app.